requirements.txt
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
cssselect==1.1.0
cssutils==2.3.0
emails==0.6
Flask==1.1.2
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.7.1
MarkupSafe==2.0.1
mysqlclient==2.1.0
premailer==3.10.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==1.0.1
wincertstore==0.2
